refactor: delete unused files
See 7c52e15f6450da257e777238dc993e96747cc8d8
