--- src/colors/palette/background.colors.ts
@@ -1,21 +1,0 @@
-export const background_100 = '#2E4759';
-export const background_200 = '#243C4D';
-export const background_300 = '#1C3242';
-export const background_400 = '#152836';
-export const background_500 = '#0F202D';
-export const background_600 = '#0B1A25';
-export const background_700 = '#08151F';
-export const background_800 = '#050F17';
-export const background_900 = '#020A0F';
-
-export default {
-  background_100,
-  background_200,
-  background_300,
-  background_400,
-  background_500,
-  background_600,
-  background_700,
-  background_800,
-  background_900
-};
