--- src/colors/palette/blue.colors.ts
@@ -1,21 +1,0 @@
-export const blue_100 = '#ebf8ff';
-export const blue_200 = '#bee3f8';
-export const blue_300 = '#90cdf4';
-export const blue_400 = '#63b3ed';
-export const blue_500 = '#4299e1';
-export const blue_600 = '#3182ce';
-export const blue_700 = '#2b6cb0';
-export const blue_800 = '#2c5282';
-export const blue_900 = '#2a4365';
-
-export default {
-  blue_100,
-  blue_200,
-  blue_300,
-  blue_400,
-  blue_500,
-  blue_600,
-  blue_700,
-  blue_800,
-  blue_900
-};
