--- src/colors/palette/gray.colors.ts
@@ -1,21 +1,0 @@
-export const gray_100 = '#f7fafc';
-export const gray_200 = '#edf2f7';
-export const gray_300 = '#e2e8f0';
-export const gray_400 = '#cbd5e0';
-export const gray_500 = '#a0aec0';
-export const gray_600 = '#718096';
-export const gray_700 = '#4a5568';
-export const gray_800 = '#2d3748';
-export const gray_900 = '#1a202c';
-
-export default {
-  gray_100,
-  gray_200,
-  gray_300,
-  gray_400,
-  gray_500,
-  gray_600,
-  gray_700,
-  gray_800,
-  gray_900
-};
