--- src/colors/palette/green.colors.ts
@@ -1,21 +1,0 @@
-export const green_100 = '#f0fff4';
-export const green_200 = '#c6f6d5';
-export const green_300 = '#9ae6b4';
-export const green_400 = '#68d391';
-export const green_500 = '#48bb78';
-export const green_600 = '#38a169';
-export const green_700 = '#2f855a';
-export const green_800 = '#276749';
-export const green_900 = '#22543d';
-
-export default {
-  green_100,
-  green_200,
-  green_300,
-  green_400,
-  green_500,
-  green_600,
-  green_700,
-  green_800,
-  green_900
-};
