--- src/colors/palette/indigo.colors.ts
@@ -1,21 +1,0 @@
-export const indigo_100 = '#ebf4ff';
-export const indigo_200 = '#c3dafe';
-export const indigo_300 = '#a3bffa';
-export const indigo_400 = '#7f9cf5';
-export const indigo_500 = '#667eea';
-export const indigo_600 = '#5a67d8';
-export const indigo_700 = '#4c51bf';
-export const indigo_800 = '#434190';
-export const indigo_900 = '#3c366b';
-
-export default {
-  indigo_100,
-  indigo_200,
-  indigo_300,
-  indigo_400,
-  indigo_500,
-  indigo_600,
-  indigo_700,
-  indigo_800,
-  indigo_900
-};
