--- src/colors/palette/orange.colors.ts
@@ -1,21 +1,0 @@
-export const orange_100 = '#fffaf0';
-export const orange_200 = '#feebc8';
-export const orange_300 = '#fbd38d';
-export const orange_400 = '#f6ad55';
-export const orange_500 = '#ed8936';
-export const orange_600 = '#dd6b20';
-export const orange_700 = '#c05621';
-export const orange_800 = '#9c4221';
-export const orange_900 = '#7b341e';
-
-export default {
-  orange_100,
-  orange_200,
-  orange_300,
-  orange_400,
-  orange_500,
-  orange_600,
-  orange_700,
-  orange_800,
-  orange_900
-};
