--- src/colors/palette/pink.colors.ts
@@ -1,21 +1,0 @@
-export const pink_100 = '#fff5f7';
-export const pink_200 = '#fed7e2';
-export const pink_300 = '#fbb6ce';
-export const pink_400 = '#f687b3';
-export const pink_500 = '#ed64a6';
-export const pink_600 = '#d53f8c';
-export const pink_700 = '#b83280';
-export const pink_800 = '#97266d';
-export const pink_900 = '#702459';
-
-export default {
-  pink_100,
-  pink_200,
-  pink_300,
-  pink_400,
-  pink_500,
-  pink_600,
-  pink_700,
-  pink_800,
-  pink_900
-};
