--- src/colors/palette/purple.colors.ts
@@ -1,21 +1,0 @@
-export const purple_100 = '#faf5ff';
-export const purple_200 = '#e9d8fd';
-export const purple_300 = '#d6bcfa';
-export const purple_400 = '#b794f4';
-export const purple_500 = '#9f7aea';
-export const purple_600 = '#805ad5';
-export const purple_700 = '#6b46c1';
-export const purple_800 = '#553c9a';
-export const purple_900 = '#44337a';
-
-export default {
-  purple_100,
-  purple_200,
-  purple_300,
-  purple_400,
-  purple_500,
-  purple_600,
-  purple_700,
-  purple_800,
-  purple_900
-};
