--- src/colors/palette/red.colors.ts
@@ -1,21 +1,0 @@
-export const red_100 = '#fff5f5';
-export const red_200 = '#fed7d7';
-export const red_300 = '#feb2b2';
-export const red_400 = '#fc8181';
-export const red_500 = '#f56565';
-export const red_600 = '#e53e3e';
-export const red_700 = '#c53030';
-export const red_800 = '#9b2c2c';
-export const red_900 = '#742a2a';
-
-export default {
-  red_100,
-  red_200,
-  red_300,
-  red_400,
-  red_500,
-  red_600,
-  red_700,
-  red_800,
-  red_900
-};
