--- src/colors/palette/teal.colors.ts
@@ -1,21 +1,0 @@
-export const teal_100 = '#e6fffa';
-export const teal_200 = '#b2f5ea';
-export const teal_300 = '#81e6d9';
-export const teal_400 = '#4fd1c5';
-export const teal_500 = '#38b2ac';
-export const teal_600 = '#319795';
-export const teal_700 = '#2c7a7b';
-export const teal_800 = '#285e61';
-export const teal_900 = '#234e52';
-
-export default {
-  teal_100,
-  teal_200,
-  teal_300,
-  teal_400,
-  teal_500,
-  teal_600,
-  teal_700,
-  teal_800,
-  teal_900
-};
