--- src/colors/palette/yellow.colors.ts
@@ -1,21 +1,0 @@
-export const yellow_100 = '#fffff0';
-export const yellow_200 = '#fefcbf';
-export const yellow_300 = '#faf089';
-export const yellow_400 = '#f6e05e';
-export const yellow_500 = '#ecc94b';
-export const yellow_600 = '#d69e2e';
-export const yellow_700 = '#b7791f';
-export const yellow_800 = '#975a16';
-export const yellow_900 = '#744210';
-
-export default {
-  yellow_100,
-  yellow_200,
-  yellow_300,
-  yellow_400,
-  yellow_500,
-  yellow_600,
-  yellow_700,
-  yellow_800,
-  yellow_900
-};
